requirements.txt
ctransformers==0.2.5
# sentence-transformers==2.2.2
sentence-transformers
pinecone-client
# langchain==0.0.225
langchain==0.2.3
flask==3.0.0
pypdf==3.17.3
langchain-pinecone==0.1.1
langchain-community==0.2.4
langchain-huggingface==0.0.3
python-dotenv
-e .